use-auth hook deleted => singIn, logOut, signUp => use-database
