+++ src/components/SignUpForm.js
@@ -1,12 +1,10 @@
 import { useState } from "react";
 import { Link } from 'react-router-dom';
-import { useAuth } from '../hooks/use-auth';
 import { useDatabase } from "../hooks/use-database";
 
 export default function SignUpForm() {
 
-    const { signUp } = useAuth();
-    const { updateUserData } = useDatabase();
+    const { signUp } = useDatabase();
 
     const [userSignUpData, setUserSignUpData] = useState({
         firstName: "",
@@ -70,7 +68,6 @@ export default function SignUpForm() {
                     className="btn btn-primary mb-3"
                     onClick={() => {
                         signUp(userSignUpData.email, userSignUpData.password);
-                        //updateUserData(userSignUpData);
                     }}
                 >
                     Create account
