+++ src/index.js
@@ -7,17 +7,14 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { BrowserRouter } from "react-router-dom";
-import { AuthProvider } from "./hooks/use-auth";
 import { DatabaseProvider } from "./hooks/use-database";
 
 ReactDOM.render(
   <React.StrictMode>
     <BrowserRouter>
-      <AuthProvider>
-        <DatabaseProvider>
-          <App />
-        </DatabaseProvider>
-      </AuthProvider>
+      <DatabaseProvider>
+        <App />
+      </DatabaseProvider>
     </BrowserRouter>
   </React.StrictMode>,
   document.getElementById('root')
